Extract filters bar component
